+++ inst/resources/activate.R
@@ -8,9 +8,11 @@ local({
   project <- getwd()
 
   # figure out path to 'renv' folder from this script
-  call <- sys.call(1L)
-  if (is.call(call) && identical(call[[1L]], as.symbol("source")))
-    Sys.setenv(RENV_PATHS_RENV = dirname(call[[2L]]))
+  if (is.na(Sys.getenv("RENV_PATHS_RENV", unset = NA))) {
+    call <- sys.call(1L)
+    if (is.call(call) && identical(call[[1L]], as.symbol("source")))
+      Sys.setenv(RENV_PATHS_RENV = dirname(call[[2L]]))
+  }
 
   # figure out whether the autoloader is enabled
   enabled <- local({