+++ tests/testthat/test-sandbox.R
@@ -4,6 +4,7 @@ context("Sandbox")
 test_that("the sandbox can be activated and deactivated", {
 
   renv_sandbox_deactivate()
+  renv_scope_options(renv.config.sandbox.enabled = TRUE)
   libpaths <- .libPaths()
   syslib <- .Library
   renv_sandbox_activate()
@@ -17,6 +18,7 @@ test_that("the sandbox can be activated and deactivated", {
 test_that("multiple attempts to activate sandbox are handled", {
 
   renv_sandbox_deactivate()
+  renv_scope_options(renv.config.sandbox.enabled = TRUE)
   libpaths <- .libPaths()
   syslib <- .Library
   renv_sandbox_activate()
@@ -29,7 +31,20 @@ test_that("multiple attempts to activate sandbox are handled", {
 
 })
 
+test_that(".Library.site isn't used even when sandbox is disabled", {
+
+  renv_sandbox_deactivate()
+  renv_scope_options(renv.config.sandbox.enabled = FALSE)
+  sitelib <- .Library.site
+  renv_sandbox_activate()
+  expect_false(sitelib %in% .libPaths())
+  renv_sandbox_deactivate()
+
+})
+
 test_that("re-activate sandbox when all is said and done", {
+  options(renv.config.sandbox.enabled = TRUE)
   renv_sandbox_activate()
   expect_false(.Library == renv_libpaths_system())
 })
+